style: study용 package에 study때 사용해본 코드 이동
