+++ app/src/androidTest/java/nextstep/signup/study/LearningComposeTest.kt
@@ -121,3 +121,4 @@ fun tests() {
         )
     )
 }
+
